Fixed change to notmuch.el space behavior.
The change is modeled after the source code for
notmuch-show-archive-message-then-next-or-next-thread (i.e., the binding for
the a key) minus the archive part of it.  If one needs this behavior as a
separate function, it should be trivial to pull out (rip off lambda, take
what's inside, bind it with a defun).