New setup.  Wicd and trayer in a new way.
Wicd is nice but untested on mschapv2.  Trayer, on the other hand, is a little
fucking bitch.  It ignores options to --height lower than about 18, so it gets
chopped off in the current setup.  (This is because it's not indicated as a
docstrut.)  The alternative is to have a gap between the bottom of xmobar and
the top of my drawable screen, which isn't there on the trayer section.
Xmobar, by the way, ALSO ignores font and size parameters.  I'm lucky I like
the way it looks right now because it's not changable.